+++ src/app/core/models/contato.model.ts
@@ -1,11 +1,18 @@
-import { Conexoes } from './../../shared/interface';
-
 export interface Contato {
   id: string | undefined ;
   nome: string;
-  sobrenome: string;
   celular: string;
   email: string;
   dataNascimento: string;
   conexoes?: Conexoes;
+}
+
+export interface Conexoes {
+    telefone: string;
+    whatsapp: string;
+    emailPadrao: string;
+    gmail: string;
+    github: string;
+    teams: string;
+    linkedin: string;
 }
