+++ src/app/lista-contatos/formulario/formulario.html
@@ -22,7 +22,7 @@
             <div class="col-6">
               <mat-form-field appearance="outline" class="w-100">
                 <mat-label>Data de Nascimento</mat-label>
-                <input matInput [matDatepicker]="picker" formControlName="dataNascimento" placeholder="dd/mm/aaaa">
+                <input readonly matInput [matDatepicker]="picker" formControlName="dataNascimento" placeholder="dd/mm/aaaa">
                 <mat-datepicker-toggle matSuffix [for]="picker"></mat-datepicker-toggle>
                 <mat-datepicker #picker></mat-datepicker>
                 <mat-error *ngIf="definirMensagemErro('dataNascimento')">
